FIX: relaxed the security measure to prevent someoene deleting the whole table

+++ ResourceStatusSystem/Utilities/MySQLWrapper.py
@@ -176,10 +176,6 @@ def delete( rssDB, params, meta ):
   params = _capitalize( params )
   params = _discardNones( params )
 
-  # Small secutiry measure, this prevents full table deletion.. by mistake I hope.
-  if not params:
-    return S_ERROR( 'Dude, you are going to delete the whole table !')
-
   # Protection to avoid misunderstandings between MySQLMonkey and new code.
   if set( meta.keys() ) - set( accepted_keys ):
     return S_ERROR( 'Delete statement only accepts %s, got %s' % ( accepted_keys, meta.keys() ) )
@@ -195,6 +191,10 @@ def delete( rssDB, params, meta ):
   field, value = None, None  
   if 'older' in meta:
     field, value = meta[ 'older' ]
+
+  # Small secutiry measure, this prevents full table deletion.. by mistake I hope.
+  if not params and not field and not value:
+    return S_ERROR( 'Dude, you are going to delete the whole table %s' % tableName )
     
   return rssDB.database.deleteEntries( tableName, condDict = params, older = value,
                                        timeStamp = field )
